Added integration tests

+++ src/OzonRoute.Api/Responses/V3/GetHistoryResponse.cs
@@ -5,6 +5,6 @@ public record GetHistoryResponse (
     long UserId,
     CargoResponse Cargo,
     double Distance,
-    double Price,
+    decimal Price,
     DateTimeOffset At
 ) {}
