+++ src/OzonRoute.Domain/Models/CalculationLogModel.cs
@@ -7,7 +7,7 @@ public record CalculationLogModel(
     DateTimeOffset At,
     double TotalVolume = 0.0, //In cm^3
     double TotalWeight = 0.0, //In kg
-    double Price = 0.0,
+    decimal Price = 0,
     double Distance = 0.0 // In metrs
 )
 { }
